+++ jk_src/jk_lib/json/lex_test.c
@@ -17,10 +17,10 @@ int main(int argc, char **argv)
     jk_platform_console_utf8_enable();
 
     JkPlatformArenaVirtualRoot arena_root;
-    JkArena storage = jk_platform_arena_virtual_init(&arena_root, (size_t)1 << 36);
+    JkArena storage = jk_platform_arena_virtual_init(&arena_root, (int64_t)1 << 36);
 
     JkBuffer text = jk_platform_file_read_full(&storage, "../jk_src/jk_lib/json/lex_test.json");
-    uint64_t pos = 0;
+    int64_t pos = 0;
 
     JkJsonToken token;
     do {
