+++ jk_src/perfaware/part3/page_fault_stats.c
@@ -34,9 +34,9 @@ JkOptionResult opt_results[OPT_COUNT] = {0};
 
 JkOptionsParseResult opts_parse = {0};
 
-int main(int argc, char **argv)
+int32_t main(int32_t argc, char **argv)
 {
-    int page_count = 0;
+    int32_t page_count = 0;
     jk_options_parse(argc, argv, opts, opt_results, OPT_COUNT, &opts_parse);
     if (opts_parse.operand_count == 1) {
         page_count = jk_parse_positive_integer(opts_parse.operands[0]);
@@ -64,28 +64,28 @@ int main(int argc, char **argv)
         exit(opts_parse.usage_error);
     }
 
-    uint64_t page_size = jk_platform_page_size();
-    JkBuffer buffer = {.size = (uint64_t)page_count * page_size};
-    for (int touch_page_count = 0; touch_page_count < page_count; touch_page_count++) {
+    int64_t page_size = jk_platform_page_size();
+    JkBuffer buffer = {.size = page_count * page_size};
+    for (int32_t touch_page_count = 0; touch_page_count < page_count; touch_page_count++) {
         buffer.data = jk_platform_memory_alloc(buffer.size);
         if (!buffer.data) {
             continue;
         }
 
         uint64_t count_before = jk_platform_page_fault_count_get();
-        uint64_t touch_size = touch_page_count * page_size;
-        for (size_t i = 0; i < touch_size; i++) {
-            size_t index = opt_results[OPT_REVERSE].present ? touch_size - 1 - i : i;
+        int64_t touch_size = touch_page_count * page_size;
+        for (int64_t i = 0; i < touch_size; i++) {
+            int64_t index = opt_results[OPT_REVERSE].present ? touch_size - 1 - i : i;
             buffer.data[index] = (uint8_t)index;
         }
         uint64_t count_after = jk_platform_page_fault_count_get();
 
-        uint64_t fault_count = count_after - count_before;
-        printf("%d, %d, %llu, %lld\n",
+        int64_t fault_count = count_after - count_before;
+        printf("%d, %d, %lld, %lld\n",
                 page_count,
                 touch_page_count,
                 (long long)fault_count,
-                (long long)(fault_count - (uint64_t)touch_page_count));
+                (long long)(fault_count - touch_page_count));
 
         jk_platform_memory_free(buffer.data, buffer.size);
     }
